requirements.txt
hdbscan==0.8.27
joblib==1.1.0
mysql-connector-python==8.0.28
numpy==1.16.3
pandas==0.24.2
scikit-learn==0.24.2
scikit-learn-extra==0.2.0
scipy==1.2.1
SQLAlchemy==1.3.20